requirements.txt
telethon
python-dotenv
cryptg